Split typechecker declaration collection
